+++ src/main/python/ayab/ayab.py
@@ -28,7 +28,7 @@ from PyQt5.QtCore import Qt, QThread, QCoreApplication, QTimer
 
 from .main_gui import Ui_MainWindow
 from .fsm import FSM
-from .observer import Observer
+from .signal_receiver import Observer
 from .audio import AudioPlayer
 from .menu import Menu
 from .scene import Scene
