+++ src/main/python/ayab/preferences.py
@@ -29,7 +29,7 @@ from PyQt5.QtCore import Qt, QSettings, QCoreApplication
 from PyQt5.QtWidgets import QDialog, QFormLayout, QLabel, QCheckBox, QComboBox
 
 from .prefs_gui import Ui_Prefs
-from .observable import Observable
+from .signal_sender import Observable
 from .engine.options import Alignment
 from .engine.mode import Mode
 from .machine import Machine
